refactor: merge the Navbar component into the Header component
