+++ src/app/components/Header/index.test.tsx
@@ -21,4 +21,57 @@ describe("<Header/> Test Suite", () => {
     const backgroundPicture = screen.getByRole("presentation");
     expect(backgroundPicture).toBeInTheDocument();
   });
+
+  it("should render the <Header/> component unchanged", () => {
+    const { container } = render(<Header />);
+    expect(container).toMatchSnapshot();
+  });
+
+  it("should contain the About link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("About");
+    expect(button).toBeInTheDocument();
+  });
+
+  it("should contain the Experience link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Experience");
+    expect(button).toBeInTheDocument();
+  });
+
+  it("should contain the Blog link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Blog");
+    expect(button).toBeInTheDocument();
+  });
+
+  it("should contain the Contact link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Contact");
+    expect(button).toBeInTheDocument();
+  });
+
+  it("should contain the correct About link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("About");
+    expect(button).toHaveAttribute("href", "/#about");
+  });
+
+  it("should contain the correct Experience link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Experience");
+    expect(button).toHaveAttribute("href", "/#experience");
+  });
+
+  it("should contain the correct Blog link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Blog");
+    expect(button).toHaveAttribute("href", "/#blog");
+  });
+
+  it("should contain the correct Contact link in the <Header/> component", async () => {
+    render(<Header />);
+    const button = screen.getByText("Contact");
+    expect(button).toHaveAttribute("href", "/#contact");
+  });
 });
