--- src/app/components/Navbar/index.module.scss
@@ -1,13 +1,0 @@
-@import "../../variables.module.scss";
-
-.Main {
-  display: flex;
-  flex-direction: row-reverse;
-  background-color: $background-color;
-  padding: 2rem;
-}
-
-.Link {
-  font-size: 1.17em;
-  padding: 0 1rem 0 1rem;
-}
